backend/requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.2

# Environment and configuration
python-dotenv>=1.0.0

# HTTP client and async support
httpx>=0.27.2
aiofiles>=24.1.0

# Retry and resilience
tenacity>=8.5.0

# Database
sqlalchemy>=2.0.35
aiosqlite>=0.20.0

# AI and ML
openai>=1.0.0
tiktoken>=0.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0

# Optional: Advanced features
# google-generativeai>=0.3.0
# anthropic>=0.7.0
# boto3>=1.28.0
# redis>=4.5.0
# celery>=5.3.0

# Development and testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
